# chore: Optimize`TraitsResolver`