### Description

There were a lot of unnecessary `Vec` allocations in `TraitsResolver`. I refactored the code to pass around iterators instead of `Vec`s, allowing the removal of several calls to `.collect()`

This gives about a **1%** improvement in block processing times:

```console
❯ hyperfine -w 3 -r 20 "target/release/stacks-inspect.next replay-block /home/jbencin/data/next/ range 99990 100000" "target/release/stacks-inspect.opt-traits-resolver-2 replay-block /home/jbencin/data/next/ range 99990 100000"
Benchmark 1: target/release/stacks-inspect.next replay-block /home/jbencin/data/next/ range 99990 100000
  Time (mean ± σ):      8.672 s ±  0.026 s    [User: 8.187 s, System: 0.433 s]
  Range (min … max):    8.618 s …  8.725 s    20 runs
 
Benchmark 2: target/release/stacks-inspect.opt-traits-resolver-2 replay-block /home/jbencin/data/next/ range 99990 100000
  Time (mean ± σ):      8.621 s ±  0.025 s    [User: 8.134 s, System: 0.441 s]
  Range (min … max):    8.578 s …  8.675 s    20 runs
 
Summary
  target/release/stacks-inspect.opt-traits-resolver-2 replay-block /home/jbencin/data/next/ range 99990 100000 ran
    1.01 ± 0.00 times faster than target/release/stacks-inspect.next replay-block /home/jbencin/data/next/ range 99990 100000
```

### Applicable issues

- #4316